Remove tag limitation and fix thread safety issues
**1. Remove RegisterTag Method (Unused)**
- Removed RegisterTag<T>() method - not used anywhere in codebase
- Kept GetOrRegisterTag which is actually used in AddComponent

**2. Fix Tag Limitation - Support Unlimited Components**
- Changed BitIndex from byte (0-255) to int in TypeBitIndex<T>
- Updated GetTagBitIndex, TryGetTagBitIndex return types: byte → int
- Updated EnsureComponentArrayInitialized parameter: byte → int
- Removed artificial "256 * 4" limit in GetOrRegisterTag
- Tag structure already supports unlimited components via overflow system

**Impact:** Framework now supports truly unlimited component types (not just 1024)

**3. Fix Critical Bug in GetOrCreateArchetype**
- Line 34: Fixed `Archetypes[componentMask] = componentMask;`
- Should be: `Archetypes[componentMask] = archetype;`
- This was storing the wrong value in the dictionary!

**4. Fix Thread Safety - Archetypes Synchronization**
**Problem:** Dictionary<Tag, Archetype> accessed both with and without locks
- Line 26: Read without lock (race condition with line 34 write)
- Dictionary is NOT thread-safe for concurrent reads during writes

**Solution:** Use ConcurrentDictionary for Archetypes
- Changed: `Dictionary<Tag, Archetype>` → `ConcurrentDictionary<Tag, Archetype>`
- Simplified GetOrCreateArchetype using GetOrAdd (thread-safe, zero allocation with static lambda)
- Removed double-check locking pattern (no longer needed)
- Clear() still works (ConcurrentDictionary has Clear method)

**Benefits:**
- Thread-safe for concurrent access from multiple systems
- No more "field sometimes used with/without synchronization" warnings
- Cleaner code with GetOrAdd instead of manual locking
- Static lambda in GetOrAdd avoids closure allocation

**Files Modified:**
- Components/TagRegistry.cs: byte → int, removed RegisterTag, removed limit
- Context.cs: Updated byte → int for component indices
- Context.Archetypes.cs: Dictionary → ConcurrentDictionary, simplified GetOrCreateArchetype

+++ EasyEcs.Core/Context.cs
@@ -264,7 +264,7 @@ public partial class Context : IAsyncDisposable
                 throw new InvalidOperationException($"Entity {entity.Id} not found.");
 
             // Get or register component type
-            byte componentIdx = TagRegistry.GetOrRegisterTag<T>();
+            int componentIdx = TagRegistry.GetOrRegisterTag<T>();
 
             // Ensure component array exists
             if (Components == null || componentIdx >= Components.Length)
@@ -346,7 +346,7 @@ public partial class Context : IAsyncDisposable
     /// Thread-safe.
     /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveOptimization)]
-    internal T[] EnsureComponentArrayInitialized<T>(byte componentIdx) where T : struct, IComponent
+    internal T[] EnsureComponentArrayInitialized<T>(int componentIdx) where T : struct, IComponent
     {
         lock (_structuralLock)
         {
